MergeTree: Unify HierBlock and MergeBlock and Simplify Usage (#20646)
The code today always uses HierBlock. This change removes the HeirBlock
class, and moves its functionality to the MergeBlock as there is no need
for a distinction between the two. Lastly this change inlines a bunch of
code only used by blockUpdate into blockUpdate. This inlining saves
codes and makes the logic easier to follow.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>
Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>