fix(build-tools): Update broken tests (#20700)
Some build-tools tests end up examining the repo state, since the
build-tools typically work with the repo and very little is easily
mockable. This does make some of the tests brittle and susceptible to
breaks due to changes elsewhere in the repo.

In this case, the common-definitions package was deleted which caused a
package filtering test to start failing, and the RC3 version number was
used in a test case but it no longer works for that purpose because we
have released that version.