Make guestDisplayname unnecessary in join session request (#20671)
## Description


[Task](https://dev.azure.com/fluidframework/internal/_workitems/edit/7614)

fetchJoinSession allows requestSocketToken argument which can be set to
true to instruct /joinSession backend to fetch access token for Push
service and return it back as part of joinSession response payload.
Current implementation sends such instruction only if guestDisplayName
argument value is not undefined. That used to be required by joinSession
implementation, but it is no longer the case. In fact,
requestSocketToken is set to true in MSA case and the caller is forced
to specify guestDisplayName which does not make sense in case of MSA
user (it is not guest scenario).

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>