fix: Add runtime error when setting via index in array nodes (#20640)
## Description

SharedTree array nodes currently attempt to prevent the setting of
values via index at both compile time and runtime. However, the runtime
behavior is mismatched - it only errors for indices which are less than
the length of the array. Furthermore, this was only a bug in array nodes
created via a _class-based_ schema, but was correct for array nodes
created via structural array schema.

This PR fixes the bug, but also establishes a pattern for testing both
the class-based and structural-based behavior of array nodes. This
pattern can be extended to the other node types in the future. The
testing has been put in a new "arrayNode.spec.ts" file to match the
current evolution of our test files (e.g. "objectNode.spec.ts").