fix(test-version-utils): Workaround for calculating N-1 version when expected release has not been published yet (#20687)
## Description

This PR changes how `getRequestedVersion()` calculates N-1 versions. If
the expected N-1 version does not exist (not published to npm yet), then
it will attempt to fetch the N-2 version in it's place.

This was done to mitigate an issue where CI would fail if we tried to
bump to a new major version without publishing the previous release
first. The following PR demonstrates CI passing if we try to bump main
from RC4 -> RC5 without publishing RC4 yet:
https://github.com/microsoft/FluidFramework/pull/20688

Note: This PR only mitigates the issue to unblock future releases. This
should not be a substitute for a refactor in the near future.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>