fix(api-markdown-documenter): Correctly handle getter/setter properties on interfaces (#20713)
Getter and setter properties are processed as properties rather than
property signatures on interfaces. The code was previously assuming that
only signatures could appear on interfaces. Updates test assets with
test cases for this.

Also refactors the pattern used for iterating child items to help detect
cases like this in the future.


[AB#7743](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7743)