refactor(telemetry-utils): Remove internal telemetry types from public API surface (#20574)
## Description


The telemetry logger usage should follow these two principals:

- ITelemetryLoggerExt should be only used by @internal

- ITelemetryBaseLogger can be used by @public, @alpha, @beta

Find the spitting point that public APIs start to use Ext instead of
Base. Change it and fix it.

This is the list of types that we have identified that we need to "hide"
(make @alpha or @internal instead of @public). It may not be exhaustive,
so if we find more types that need changes, we should address them too.

- **TelemetryEventCategory**

- **TelemetryEventPropertyTypeExt**

- **ITelemetryPropertiesExt**

- **ITelemetryGenericEventExt**

- **ITelemetryErrorEventExt**

- **ITelemetryPerformanceEventExt**

- **ITelemetryLoggerExt**

- ~~ITelemetryContext~~

---------

Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>
Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>