fix(container-runtime): only send OpRoundtripTime for ops, not noops (#20701)
[AB#7583](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7583)

## Description

This PR investigates and fixes a bug in which some OpRoundtripTime
events are sent without any performance data
(`durationInboundToProcessing`, `durationOutbound` and
`durationNetwork`). This happened because the events were being sent on
noop messages as well as regular ops.

This resulted in gaps in the Op Latency view which causes gaps in the
graph.

This fix also removes the need to check for ops with the performance
data.