[Data Virtualization & Offline]: Reimplement groupId offline with the loader (#20565)
Original PR:
https://github.com/microsoft/FluidFramework/pull/20342/files


[AB#7244](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7244)

[AB#7245](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7245)

[AB#7588](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7588)

# Description
For Data virtualization and offline, we need to build a solution that is
robust when going offline and refreshing. The method we use here is to
store any snapshots that we need in order to process the offline ops.
The goal here is to provide basic Data Virtualization for offline.

## Storing
We need to store these snapshots separately when the snapshot was
downloaded at some point between the base/refreshed snapshot and before
the sequence number we went offline. This includes when we load from
pending state as we can consider that "downloading".

The best way to do this is to put it in the pending state.

## Refresh
When we refresh the state of the snapshot, we need to make sure that all
groupIds (virtualized snapshots that were not in the base snapshot) are
downloaded so that the new snapshot is up to date.
 
## Loading
To load from these snapshots, we store them separately as pending state,
so when we are apply stashed ops for offline, the right blobs are
downloaded. We don't want to download any blobs from the future when
applying stashed ops as that would lead to inconsistent/incorrect state.