GC: Properly notify GC when a custom sub-DataStore route is requested (#20725)
Key logic changes:
* GC's `nodeUpdated` function only accepts Blob or DataStore paths (no
subDataStore).
* This is because subDataStore paths may not be tracked/understood by
GC. So we use the DataStore path to check Tombstone/Inactive status.
* Note that the original subDataStore path is available for logging
since the original `request` is passed in
* `GCTelemetryTracker.nodeUsed` takes in a `trackedId` param, separate
from the props.
  * `trackedId` would be the DataStore ID in a SubDataStore case
  * It's used to look up the UnreferencedNodeStateTracker

There is some cleanup type stuff in the diff as well, in the areas I was
working. Very light refactoring and comments mostly.

I added tests to request inactive or Tombstoned SubDataStore paths, and
checked the logs to see that we get the logs we are currently missing in
these cases.