Deep Freeze Mock Messages (#20628)
We are seeing bugs where object references are being passed on ops which
then leads to corruptions when those object are modified later. This
manifests more strongly in out tests, where the object survive the op
processing round trip, but this could still cause real problems on a
local client. This change add a deep freeze on the object passed into
the op processing pipeline such that we can catch if any of the
referenced objects are later modified.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>