[main > Release/Client/rc2]: Use tenant domain instead of siteUrl in /shares api in ODSP driver (#20738) (#20770)
## Description

[Task
Link](https://dev.azure.com/fluidframework/internal/_workitems/edit/7650)

To load a Fluid container, host/app provides a URL which is used to talk
to service to load the container. That URL can be a share link which
host/app specifies explicitly to Fluid that the URL is a share link.
Fluid then uses that link to fetch the snapshot for the container. User
sometimes don’t have permissions initially, so to save COGS, Fluid
supply this share link in body of trees latest call to the service, so
that in 1 network call, we can do the redeem and also fetch the
snapshot. Sometimes this redeem on service side fails (in case it is
first time sharing with that user), then the service throws on the trees
latest call. Fluid catches this exception and determines/guess if this
is because of redeem failure at service side and then fallback to doing
redeem by itself using the /shares API. Once the redeem fails even in
/shares API, the container load fails.

In this issue, for outlook Win32 the /shares API (Documentation link for
API:

[Link](https://docs.microsoft.com/en-us/onedrive/developer/rest-api/api/shares_get))
call is failing because of the length limitation on the API. There are 2
ways to call the /shares API(for Fluid use),

Method 1: site-url/_api/v2.0/shares/{shareID}
This has length limitation of about 400 characters.

Method 2: tenant-domain/_api/v2.0/shares/{shareID} Then the length
limitation is about 2048 characters. where this, share ID is generated
by encoding the share link supplied by the app in the above load
request. So, encoding longer URL will give longer shareID.

Tenant-domain can be extracted from site-url by doing: Const
tenant-domain = new URL(site-url).origin

This issue occurred for Outlook win 32, because MDO team started adding
some, additional long query params to the share link. So, we are
switching to use Method 1 instead of Method 2.