Use execFile over exec for npm commands in versionUtils.ts (#20684)
## Description

The FF repository has a few alerts for [unsafe shell commands
constructed for user
input](https://codeql.github.com/codeql-query-help/javascript/js-shell-command-constructed-from-input/).

These alerts are all in test-only code or CLI tooling, so privilege
escalation isn't a huge risk here. However, it wasn't much effort to fix
and it's better to set a good example, so this PR does that for
test-version-utils by converting `exec` usages to `execFile` and
providing user input as arguments. This means we're no longer spawning a
shell to run these commands, as `npm` is invoked directly.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>