fix(server): give IWebSocket.emitToRoom void return type. (#20735)
## Description

I noticed when testing in local server we consistently get this error
when disconnecting from document:

'TypeError: Cannot read properties of undefined (reading 'catch')' 

This is because in disconnectDocument we do
`socket.emitToRoom().catch((error) => {...});`

There is an inconsistency in the way we type IWebSocket.emitToRoom. We
label it as `async` in some class implementations, but returning a
promise is unnecessary given its usage.


https://github.com/microsoft/FluidFramework/blob/17570a472184e0874abf49437e79810414c7670f/server/routerlicious/packages/lambdas/src/nexus/disconnect.ts#L110-L112

This PR gives IWebSocket.emitToRoom a void return type, removes async
labels in given implementations, and replaces .catch usage with
try/catch blocks instead.