Revert: Reset joinSession cache only on connect_document_error (#20784) (#20789)
Reverts https://github.com/microsoft/FluidFramework/pull/19974 PR which
was applied as a performance improvement. In the original PR, we checked
for the kind of error received on the socket to decide whether to clear
joinSession cache or not and reuse the existing session whenever
possible. However a bug was identified where one section of the code was
not update to stamp the errorFrom property which is used to perform such
check which resulted in the cache never clearing for and the code
getting stuck in a loop to continue connection retries.

Impact of revert: There is no compatibility-issue/regression that this
revert would introduce, as it was only a perf improvement.

Follow up: Captured in

[AB#7833](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7833)