Stress: Don't log errors after dispose (#20800)
We see occasional errors in stress that appear to be cascades.
Specifically, some error, which is real, causes the container to close,
but there is outstanding work in the stress test. Due to the container
closing that outstanding work can also fail, which leads to additional
errors being logged, which creates noise in the stress run. This change
simply doesn't log error if the container is closed and the runtime is
disposed. this should reduce noise in our stress runs.

Co-authored-by: Tony Murphy <anthonm@microsoft.com>