Stress cleanup disposed handling (#20807)
This change improves the handling of the disposed state of the runtime
in the stress test. Previously a runtime disposal would result in error
being thrown from the test and logged into the test harness. This leads
to unnecessary noise in the stress logs as if the container is disposed
there was likely already an error logged. Additionally, we improve how
the runner manages the container such that all faults have errors, which
the test knows now to ignore, and all closures expect an error except
for the final disposal which is the only expected, non-fault,
close/dispose.

fixes
[AB#7811](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7811)

Co-authored-by: Tony Murphy <anthonm@microsoft.com>