Update conditions to refresh the snapshot and sessionExpiryTimer (#20427)
This PR makes several fixes related to snapshot refresh:
1. When the snapshot refreshes at loading when using a pending local
state, we need to move the session expiry timer forward to avoid closing
the container sooner. The new time should be the moment in which we
fetched the new snapshot and should only override garbage collector
expiry time if we updated serializedStateManager snapshot and once the
container is saved, since we could carry ops that haven't been acked in
subsequent rehydration.
2. After solving above task, I realized we could carry saved ops that
were already processed in subsequent stashes. That happened because we
didn't clean pending state manager saved ops and we didn't check
container state before updating the snapshot. Now we are only going to
carry ops older than our latest snapshot and wait until the container is
saved to update it
3. There is an edge case scenario in which we could have fetched the
snapshot before the runtime is fully instantiated, so I added an extra
condition to update the snapshot checking that all saved ops have been
processed. This will help to avoid capturing the state again in an
inconsistent state while we load.
4. These conditions will also be needed once we periodically refresh the
snapshot.