fix(generate:entrypoints): Enable jsx tsconfig option in ts-morph project (#20780)
We don't load the whole tsconfig in ts-morph when we generate
entrypoints. Instead, we set explicit compiler options when we create
the ts-morph project. The drawback is that some projects require
settings like `jsx` and without these settings, ts-morph does not
identify exports properly.

Since `jsx` is the only known setting that we need beyond the `module`
setting, I just added it as a hardcoded setting. If we find more
settings are needed then we can pursue a bigger change.