feat(build-cli): Add generate:packlist command (#20723)
Adds a new flub command, `generate:packlist`, which can be used to
generate a file with the list of the _packed_ contents of a package.
Because some of the packed files might be built, this command should
only be run on packages that have been built. When we integrate this
into the build, we may make it depend on build, test, and webpack, since
that's what we run in CI before pack.