Add createSharedObjectKind utility (#20834)
## Description

Each DDS exposes a SharedObjectKind which does the same thing. This
utility makes doing so easier as well as centralizing the logic to make
changes to it easier.

This was done as a cleanup when considering applying ErasedType to
IChannelFactory to try and reduce the use of IChannel (and its summary
related types it exposes) and IFluidDataStoreRuntime (and the delta
types it exposes). Expirments like that will be easier with the
refactoring to use createSharedObjectKind included in this change.

Also removes a `@sealed` on a SharedObjectKind instance, since its a
variable not a class (or even interface) it doesn't make sense.