tree: Remove support for implicit revision tags (#20756)
## Description

Previously in many contexts in changesets an undefined revision tag was
interpreted as implicitly having the value of the revision tag tagged on
the changeset. This PR beings getting rid of this convention by inlining
revision tags into all changesets. Much of the logic supporting implicit
revision tags will be removed in a future PR.