Added support for custom tenant key generators (#20844)
## Description

This PR adds support for customizations for generating `tenantKeys`.
This would allow customized ways of creating `tenantKeys`. This is
needed to migrate to HIS credentials for Azure. On the FRS side, we
would need to implement the `ITenantKeyGenerator` interface with our
tenantKey genereation logic.

The change was tested on a dev cluster.

## Breaking Changes

This change breaks backward compatibility for `RiddlerResources` as it
introduces a new arg for `ITenantKeyGenerator`.