Stress: Fix detection of container closing without error (#20855)
The previous attempt to detect closes without error did not work due to
the way close and dispose are intertwined for events and states. This
change avoids that complexity by capturing the error in the closed
event, which will also be called in the case of dispose, and checking
that at the end of the loop interaction before the final dispose. If the
container is close, which also includes disposed, the error should
exist, as the following line is the only place we close/dispose without
error

Co-authored-by: Tony Murphy <anthonm@microsoft.com>