fix(build-cli): PackageCommands should error if any of the child processes fail (#20878)
The PackageCommand base class was exiting with no errors even when child
processes failed. This is especially problematic when running a
PackageCommand with the `--dir .` flag to run on a single package. If
that single package failed, the command still succeeded.

The implementation now collects and returns error strings from
subprocesses, and if any errors are returned the command will exit with
a non-zero exit code.

This should fix several problems people have experienced with commands
based on PackageCommand where builds report success but fail downstream
during testing or otherwise.