fix(generate:typetests): Load source instead of type declarations for current package version (#20885)
Fixes the build ordering problem introduced by unifying the code paths
for loading files to extract types. I added a lengthy comment explaining
why the code is different and its limitations. This is essentially
reverting the code to what it is in the current type test generator. I
had refactored it when porting it to build-cli, so I introduced the bug.

I also noticed that oclif apparently still returns some sort of error
when you call exit from a command, so I used an early return instead.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>