fix(fluid-build): Always consider semantic errors in incremental tsc (#20887)
fluid-build was not considering semantic errors as failures in some
cases. This caused fluid-build to skip tasks that were in fact not up to
date. I fixed that flaw and also made the boolean logic more explicit.
It's more verbose, but I think it clarifies the cases. Had the code been
written this way originally, I don't think the bug would have been
written.