`runTransaction` returns the value of the inner transaction function  (#20872)
This allows data to be calculated during a transaction and easily
retrieved at the call site. This also necessitated changing the special
"rollback" value (which is sometimes returned from a transaction
function to indicated the transaction should be cancelled) from a string
to a unique symbol to avoid ever colliding with a user's return value.
This ends up producing some new overloads of the runTransaction
function.

Additionally, the tests for runTransaction have been deduplicated as
much as possible to put a stop to the current doubling of all top-level
transaction tests. The tests want to test both the "pass in the
TreeView" and the "pass in the node" flavors of the API, but the tests
for each are often identical except for that detail. There is now a
helper called `runCommonTransactionTests` which holds all of these kinds
of tests. The test layout now resembles the following:


![image](https://github.com/microsoft/FluidFramework/assets/78610362/e1847176-13a1-4944-bf07-84bf722f6499)