build: Update type tests for RC3 (#20776)
Updates the type tests previous versions to RC3. There were several
packages that still have broken tests, but rather than try to review
those here, I recommend we merge this and I will follow up with a
separate PR that resets all broken tests and review that PR with package
owners to verify the breaks are legitimate.

This PR does bump the previous versions, so it is beneficial to merge
as-is rather than waiting for more changes from specific projects.