server: Limit max length of validParentSummaries (#20850)
## Description

If a document has a lot of service summaries but no client summaries,
the list of validParentSummaries grows extremely long. Older service
summaries become more unlikely to be used as parent summaries as more
summaries are generated, so we can just shift the list by 1 when the
length is too long.

To fix old docs loaded from a large checkpoint, this will remove
elements down to the limit when the gap is greater than 1.

## Reviewer Guidance

The configuration is not bubbled up to service level configs yet. I
don't think it's necessary.