feat(azure-end-to-end-tests): Setup basic smoke tests for SharedTree with AzureClient (#20925)
## Description

This PR adds very basic smoke tests for using SharedTree with
AzureClient. The intention is to create a starting point that can be
expanded upon by feature owners.

## Reviewer Guidance

Disclaimer: I am new to SharedTree and this is my first attempt at
creating basic smoke tests for SharedTree. The goal of these tests is to
expose any changes that immediately breaks SharedTree functionality with
AzureClient.