Op/Noop telemetry (#20727)
[AB#7640](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7640)

Adding some telemetry around observed ops and sent no-ops with following
goals in mind:
1. To be able to get average figures on how many ops are sent in a
session. (This data would have been useful when FRS reported the issue
of large number of ops)
2. To get an estimate of how many no-ops are in ratio of ops-processed
in the system. This data will be useful to have if/when we want to
update the no-op heuristics.
3. Have a way to estimate the average collab-window that a client
maintains on it's end. This too could be useful if/when we update no-op
heuristics with content size as one of the control parameters.

In this PR, I am connectionTelemetry utility to observe and processed op
counts, no-op counts and the size of the content processed within the
sample size logged.

In future `OpStats` event can be further updated to observe other kinds
of ops as well/collab-window size related params as well, that might
help us refine our heuristics to keep collab window smaller.