Merge Tree: Fix annotate eventual consistency issues. (#20775)
This change fixes a number of eventual consistency issues related to
annotates, and enable annotation across all relevant tests.

###  Fix prop undefined handling
Previously undefined annotate properties were not eventually consistent,
as they did not survive serialization. We explored a few options:

1. Block undefined usages.
2. Ensure all undefined get remapped to null in all cases.
3. Ignore undefined.

The worry with number 1 is it's too strong and could break customers.
That said, anyone using this is already broken, but there also may not
care, as it may be being done unintentionally, like via object
expansion.

Number 2 is really hard due to our layering. And like the above, could
cause problems in the unintentional use case.

I didn't like number 3 a first, but it turned out to be the best option,
it will break someone who is intentionally passing undefined to delete,
but this doesn't work anyway. This also seems to better match java
script guidance to not differentiate between key existence or not when
undefined.


[AB#7805](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7805)
[AB#7838](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7838)

### Don't emit sequence delta events for removed segments
Previously we would erroneously emit sequence delta events for removed
segments, which was something we had a bug to fix. However, it also
turns out this could cause eventually consistency issue with how we
compute the catch up ops stored in the summary. specifically, we were
erroneously generating catch up ops for removed segments, which would
then get applied improperly when the summary was loaded.

### Enable annotates in all relevant fuzz tests
With the above issued fixed it was possible to enable annotates across
all our fuzz tests, and remove the annotate specific fuzz, which only
existed for stabilization purposes.

[AB#7806](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7806)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>
Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>