improvement(flub list): Support independent packages (#20942)
`flub list` now supports independent packages within the repo. You can
now pass a package name or release group name as an argument to `flub
list` and it will handle it appropriately. The `--releaseGroup` flag is
now deprecated but is still present so as not to break existing CI
pipelines where we still use it.

This command is intended for use in CI, where we use the package lists
to determine what packages to publish to each npm feed. With this new
capability we can use package lists for individual packages, thus
unifying the way the CI pipeline works between release groups and
independent packages.

In addition, if the path to the output file does not exist, it will be
created.