Upload Dev Manifest Files to Azure Blob Storage (#20722)
This PR uploads dev manifest files to azure blob. The following bunch of
steps are added to the pipeline:

1. Run `flub release report -g client -o <path_generate_manifest_files>
--baseFileName foo` - This command generates the manifest file as per
the current state of the repository. It names the manifest files as
`foo.full.json, foo.caret.json, foo.simple.json, foo.tilde.json`

2. Run `flub release report-unreleased -V <dev_version> -c
<path_caret_manifest> -s <path_simple_manifest> -o <output_directory>` -
This command accepts the dev version that is set in the pipeline. Later,
it updates the generated manifest files in the pipeline pointing the
values to the dev version. The manifest files undergo renaming in two
formats: one based on the date and the other on the build number.

3. Later, after uploading the dev manifest files to azure blob, the
manifest files are deleted to avoid failures at the step `Check for
extraneous modified files`.

Pipeline Run:
https://dev.azure.com/fluidframework/internal/_build/results?buildId=259575&view=logs&s=6884a131-87da-5381-61f3-d7acc3b91d76

This is how the upload folder looks:

![image](https://github.com/microsoft/FluidFramework/assets/48232592/22d3d277-916a-44c4-bc6a-8257b35cacd1)



[AB#7884](https://dev.azure.com/fluidframework/internal/_workitems/edit/7884)