Align telemetryContextPrefix for SharedTree (#20418)
https://github.com/microsoft/FluidFramework/pull/10294 added the
capability to collect some summarize telemetry from DDSes. Whenever this
`SharedTreeCore` object was created it used a different convention than
the other DDSes.

Ideally the `telemetryContextPrefix` wouldn't need to be present in the
ctor, but that's a breaking change for these classes.

Note that `"fluid_sharedTree_"` is currently used in
[`experimental/.../SharedTree.ts`](https://github.com/microsoft/FluidFramework/blob/main/experimental/dds/tree/src/SharedTree.ts#L544).
Querying on `Data_runtimeVersion` will be needed as it will now be using
`"fluid_legacySharedTree_"`.