Removal of omitted flag from snapshot as service will not supply it. (#20507)
## Description

Service will not add the omitted flag to tree structure if the blobs are
missing. Datastore will ask container runtime whether to fetch the
snapshot or not before using it.
Remove omitted prop on ISnapshotTree as service will not supply it. It
was added to be used in Data Virtualization but it was never used or
supplied by the service.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>