Exclude ChunkedOp from LocalContainerRuntimeMessage  (#20981)
We try to model the valid types of ContaineRuntime Messages for various
scenarios - `LocalContainerRuntimeMessage` models ops that can be
generated from local interaction before the Outbox does op
virtualization. `ChunkedOp` is not a valid type at that layer, so
exclude it from the type and remove handling of it from the
corresponding switch statements.