fix(matrix): Convert reentrancy detection to a usage error (#20992)
## Description

This error should not be an assert, as it validates applications don't
make matrix edits in response to matrix changes.

I also added a try-finally on the validation, as otherwise `callback`
can throw an error and result in the method permanently throwing
whenever it's called. This was the case for some telemetry I looked at
recently, where an application did *not* have reentrant code, but they
*did* have some improperly disposed processes which attempted to submit
ops on a closed container. The first one failed with a reasonable error
(validation against "runtime is closed"), but subsequent ones claimed
reentrancy.

---------

Co-authored-by: Abram Sanderson <absander@microsoft.com>