fix(generate:typetests): Use types/typings field only for public exports when exports map is not defined (#20989)
If a package has an export map, then the types/typings field should be
ignored. In addition, those fields should be ignored when looking for an
export matching a non-public API level.

Fixing these bugs also fixes other fallback logic that was being skipped
because the type/typings field was checked too early and returned those
values for non-public entrypoints like internal. Now, the logic
correctly ignores types/typings for the internal entrypoint, allowing
the fallback logic to kick in and try the public entrypoint. At _that_
point, the types/typings field will be used if and only if the package
lacks an export map.