tree: Update typebox (#20719)
## Description

TypeBox is ~50 releases behind, and this updates it. This update is
required to support TypeScript 5.4 and thus the new API-extractor
release.

While evaluating this, I determined that the typebox compiler is not
getting tree shaken correctly, resulting in unexpectedly large bundles
for it. The compiler size increased a bit between 0.30.4 and 0.31.0
(typebox added better errors with internalization support and several
other features), and thus the inclusion of the compiler results in a
regression in bundle size.

There is also a smaller regression (~3kb) from 0.31.0 to the latest's
0.32.29 even when excluding the compiler.

Investigating the tree shaking issue is being tracked by
https://dev.azure.com/fluidframework/internal/_workitems/edit/7995