Partner Ask: Re-add Removed Exports (#21042)
We're received feedback that our current RC releases are missing exports
one of our partners depends on. Some of these are legitimate misses,
where other represent coupling we do not want. The bad couplings have
been marked as deprecated, and we have a plan with the partner to move
off those types, but we need this change to unblock integrations.

related to
[AB#8004](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8004)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>