fix(id-compressor): Prevent resubmission of ID allocation ops (#21043)
This PR makes the runtime correctly resubmit ID allocation ops;
specifically, prior to initiating the replay of pending ops the runtime
will submit an allocation range that includes all pending IDs and then
ignore allocation ops in resubmit. This is necessary because the
resubmission of non-allocation ops can cause allocations (which would
then be resubmitted before earlier allocation ops, causing the "Ranges
finalized out of order." exception).

This PR also differentiates exception messages in the compressor for
this failure case.