User summarylogger when emitting summarize telemetry (#21064)
As a part of investigating
[AB#7652](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7652)

When trying to link `summarize_end` and `SummarizeTelemetry` events we
found via the `summaryNumber`, we could not because the
`SummarizeTelemetry` event comes from the root logger. We should use the
`summaryLogger`/`summaryNumberLogger` so that the `summaryNumber` and
`containerId` can be used to correlate the different summaries together.